The freeze: an unsigned remainder minus a constant, and a line that never ended
Every freeze today was the same frame. The scenery's prop scatter
computed a position as an unsigned remainder minus a hundred, and in
C# an int constant slides into unsigned arithmetic, so about one prop
in fourteen wrapped to four billion, which the int cast saturated to
the far edge of the integers. A prop drawn with a line — the coast's
palms, the highland's tufts, the void's spars — then sent the
Bresenham walk off toward a point it could never reach, and the
render thread spun until the crash handler killed the game. Forest
props are discs and fills, which is why the forest never froze and
why the harness, which only knew the forest, saw nothing.

The arithmetic is signed now, and the line drawer stops after twice
the canvas's perimeter no matter what it is handed. The harness
sweeps every biome and every story plate clean. The unload step
logging stays; it was how the stack got read.

+++ src/Aetherstream.Plugin/Video/Scenery.cs
@@ -109,7 +109,10 @@ internal static class Scenery
         for (var i = 0; i < 9; i++)
         {
             prng ^= prng << 13; prng ^= prng >> 17; prng ^= prng << 5;
-            var px = (int)((prng % (uint)(width + 200)) - 100 - (scroll * 0.9) % (width + 200));
+            // Signed on purpose: an unsigned remainder minus a constant stays unsigned and wraps
+            // to four billion, which the int cast then saturates — and a prop at the far edge of
+            // the integers sends the line drawer off for a walk it never comes back from.
+            var px = (int)(prng % (uint)(width + 200)) - 100 - (int)((scroll * 0.9) % (width + 200));
             if (px < -100)
                 px += width + 200;
             var scale = 1 + (int)((prng >> 8) % 3);
